avoid conversion to float

+++ src/esp_loader.c
@@ -85,7 +85,7 @@ static inline void md5_final(uint8_t digets[16]) { }
 
 static uint32_t timeout_per_mb(uint32_t size_bytes, uint32_t time_per_mb)
 {
-    uint32_t timeout = time_per_mb * (size_bytes / 1e6);
+    uint32_t timeout = time_per_mb * (size_bytes / (uint32_t)1e6f);
     return MAX(timeout, DEFAULT_FLASH_TIMEOUT);
 }
